config: the model-override foundation moves out of overrides.py
The per-model declarations are about to leave overrides.py for one module
per family. What they are written against -- the registry they register
into, the read-only views they are handed, the accessors that answer
questions about the model -- goes first, into model_override_base.

The point is the direction of the dependency: a family module imports the
base and never has to import its way back up to overrides.py, so there is
no cycle to work around and no import-at-the-bottom trick. overrides.py
re-exports the same names, so every existing importer is untouched.

Pure move: the registry (55 architectures, 61 registrations) compares
identical, and the 62-shape resolution probe is unchanged.
